MaterialModelsBase

Documentation for MaterialModelsBase.

MaterialModelsBase.NoLocalConvergence

MaterialModelsBase.NoStressConvergence

MaterialModelsBase.get_cache

MaterialModelsBase.initial_material_state

MaterialModelsBase.material_response

MaterialModelsBase.NoLocalConvergence — Type

NoLocalConvergence(msg::String)

Throw if the material_response routine doesn't converge internally

MaterialModelsBase.NoStressConvergence — Type

NoStressConvergence(msg::String)

Throw if the stress iterations don't converge

MaterialModelsBase.get_cache — Method

get_cache(m::AbstractMaterial)

Return a cache that can be used when calling the material to reduce allocations TODO: Rename to allocate_cache?

MaterialModelsBase.initial_material_state — Method

get_initial_state(m::AbstractMaterial)

Return the (default) initial state of the material m

MaterialModelsBase.material_response — Function

material_response(m::AbstractMaterial, ϵ::SymmetricTensor{2}, old::AbstractMaterialState, Δt, cache::AbstractMaterialCache, extras::AbstractExtraOutput; options)

material_response(m::AbstractMaterial, F::Tensor{2}, old::AbstractMaterialState, Δt, cache::AbstractMaterialCache, extras::AbstractExtraOutput; options)

material_response(m::AbstractMaterial, u::Vec, old::AbstractMaterialState, Δt, cache::AbstractMaterialCache, extras::AbstractExtraOutput; options)

Calculate the stress/traction, stiffness and updated state variables for the material m, given the strain input ϵ, the deformation gradient F, or the deformation vector u

Mandatory arguments

m: Defines the material and its parameters

The second input can be one of

ϵ::SymmetricTensor{2}: The total small strain tensor at the end of the increment

F::Tensor{2}: The total deformation gradient at the end of the increment

u::Vec: The deformation at the end of the increment (for cohesive elements)

old::AbstractMaterialState: The material state variables at the end of the last converged increment The material initial material state can be obtained by the initial_material_state function.

Optional positional arguments

Δt: The time step in the current increment. The default may depend on the material.

cache::AbstractMaterialCache: Cache variables that can be used to avoid allocations during each call to the material_response function This can be created by the get_cache function.

extras: Updated with requested extra output. Defaults to the empty struct NoExtraOutput

Optional keyword arguments

options: Additional options that may be specific for each material. Defaults to nothing

Output

If the second input is the deformation gradient $\boldsymbol{F}$ (F::Tensor{2})`, the outputs are

P::Tensor{2}: First Piola-Kirchhoff stress, $\boldsymbol{P}$

dPdF::Tensor{4}: Algorithmic tangent stiffness tensor, $\mathrm{d}\boldsymbol{P}/\mathrm{d}\boldsymbol{F}$

state::AbstractMaterialState: The updated material state variables at the end of the time increment

If the second input is the small strain tensor, $\boldsymbol{\epsilon}$ (ϵ::SymmetricTensor{2}), the outputs are

σ::SymmetricTensor{2}: Cauchy stress tensor, $\boldsymbol{\sigma}$

dσdϵ::SymmetricTensor{4}: Algorithmic tangent stiffness tensor, $\mathrm{d}\boldsymbol{\sigma}/\mathrm{d}\boldsymbol{\epsilon}$

state::AbstractMaterialState: The updated material state variables at the end of the time increment

If the second input is deformation vector, $\boldsymbol{u}$ (u::Vec), the outputs are

t::Vec: Traction vector, $\boldsymbol{t}$

dtdu::SecondOrderTensor: Algorithmic tangent stiffness tensor, $\mathrm{d}\boldsymbol{t}/\mathrm{d}\boldsymbol{u}$

state::AbstractMaterialState: The updated material state variables at the end of the time increment