Conversion

MaterialModelsBase defines an interface for converting parameters and state variables to and from AbstractVectors. This is useful when doing parameter identification, or when interfacing with different languages that require information to be passed as arrays.

These function can be divided into those providing information about the material and state variables, and those doing the actual conversion.

Information functions

MaterialModelsBase.get_tensorbase — Function

get_tensorbase(m::AbstractMaterial)

Return the type of the primary input (strain-like) and associated output (stress-like) to the material model. The default is SymmetricTensor{2, 3} for small-strain material models in 3D, but it can be

Small strain material model: Small strain tensor, ϵ::SymmetricTensor{2, dim}

Large strain material model: Deformation gradient, F::Tensor{2, dim}

Traction-separation law: Displacement jump, Δu::Vec{dim}

Where dim = 3 is most common, but any value, 1, 2, or 3, is valid.

MaterialModelsBase.get_num_statevars — Function

get_num_statevars(s::AbstractMaterialState)
get_num_statevars(m::AbstractMaterial)

Return the number of state variables. A tensorial state variable should be counted by how many components it has. E.g. if a state consists of one scalar and one symmetric 2nd order tensor in 3d, get_num_statevars should return 7.

It suffices to define for the state, s, only, but defining for material, m, directly as well can improve performance.

MaterialModelsBase.get_statevar_eltype — Function

get_statevar_eltype(s::AbstractMaterialState)

Get the type used to store each scalar component of the material state variables, defaults to Float64.

MaterialModelsBase.get_num_params — Function

get_num_params(m::AbstractMaterial)

Return the number of material parameters in m. No default value implemented.

MaterialModelsBase.get_params_eltype — Function

get_params_eltype(m::AbstractMaterial)

Return the number type for the scalar material parameters, defaults to Float64

MaterialModelsBase.get_num_tensorcomponents — Function

get_num_tensorcomponents(::AbstractMaterial)

Returns the number of independent components for the given material.

Note

It is not required to implement this function, it is inferred by the implementation of get_tensorbase

Conversion functions

MaterialModelsBase.tovector! — Function

tovector!(v::AbstractVector, m::AbstractMaterial; offset = 0)

Put the material parameters of m into the vector v. This is typically used when the parameters should be fitted.

tovector!(v::AbstractVector, s::AbstractMaterialState; offset = 0)

Put the state variables in s into the vector v. This is typically used when differentiating the material wrt. the the old state variables.

tovector!(v::AbstractVector, a::Union{SecondOrderTensor, FourthOrderTensor}; offset = 0)

Puts the Mandel components of a into the vector v.

MaterialModelsBase.fromvector — Function

fromvector(v::AbstractVector, ::MT; offset = 0) where {MT<:AbstractMaterial}

Create a material of type MT with the parameters according to v

fromvector(v::AbstractVector, ::ST; offset = 0) where {ST<:AbstractMaterialState}

Create a material state of type ST with the values according to v

fromvector(v, ::TT; offset = 0) where {TT <: Union{SecondOrderTensor, FourthOrderTensor}}

Create a tensor with shape of TT with entries from the Mandel components in v.

MaterialModelsBase.tovector — Function

tovector(m::AbstractMaterial)

Out-of place version of tovector!. Relies on get_num_params and get_params_eltype to be correctly defined

tovector(m::AbstractMaterialState)

Out-of place version of tovector!. Relies on get_num_statevars and get_statevar_eltype to be correctly defined