Internal documentation

The following documentation returns to internals of the package and may change at any time.

MaterialModelsBase.stress_controlled_indices — Function

stress_controlled_indices(stress_state::AbstractStressState, ::AbstractTensor)::SVector{N, Int}

Get the N indices that are stress-controlled in stress_state. The tensor input is used to determine if a symmetric or full tensor is used.

MaterialModelsBase.strain_controlled_indices — Function

strain_controlled_indices(stress_state::AbstractStressState, ::AbstractTensor)::SVector{N, Int}

Get the N indices that are strain-controlled in stress_state. The tensor input is used to determine if a symmetric or full tensor is used.