MaterialModelsBase

Documentation for MaterialModelsBase.

The main purpose of this package is to provide a unifying interface, facilitating the interchanging of material models between researchers. The standard use case is described below. In addition, the following extra features are included

Stress iterations

Differentiation wrt. material parameters

Standard interface

The main function is the material_response function that primarly dispatches on the AbstractMaterial input type.

MaterialModelsBase.material_response — Function

material_response(m::AbstractMaterial, ϵ::SymmetricTensor{2}, old::AbstractMaterialState, Δt, cache::AbstractMaterialCache, extras::AbstractExtraOutput; options)

material_response(m::AbstractMaterial, F::Tensor{2}, old::AbstractMaterialState, Δt, cache::AbstractMaterialCache, extras::AbstractExtraOutput; options)

material_response(m::AbstractMaterial, u::Vec, old::AbstractMaterialState, Δt, cache::AbstractMaterialCache, extras::AbstractExtraOutput; options)

Calculate the stress/traction, stiffness and updated state variables for the material m, given the strain input ϵ, the deformation gradient F, or the deformation vector u

Mandatory arguments

m: Defines the material and its parameters

The second input can be one of

ϵ::SymmetricTensor{2}: The total small strain tensor at the end of the increment

F::Tensor{2}: The total deformation gradient at the end of the increment

u::Vec: The deformation at the end of the increment (for cohesive elements)

old::AbstractMaterialState: The material state variables at the end of the last converged increment The material initial material state can be obtained by the initial_material_state function.

Optional positional arguments

Δt: The time step in the current increment. The default may depend on the material.

cache::AbstractMaterialCache: Cache variables that can be used to avoid allocations during each call to the material_response function This can be created by the get_cache function.

extras: Updated with requested extra output. Defaults to the empty struct NoExtraOutput

Optional keyword arguments

options: Additional options that may be specific for each material. Defaults to nothing

Output

If the second input is the deformation gradient $\boldsymbol{F}$ (F::Tensor{2})`, the outputs are

P::Tensor{2}: First Piola-Kirchhoff stress, $\boldsymbol{P}$

dPdF::Tensor{4}: Algorithmic tangent stiffness tensor, $\mathrm{d}\boldsymbol{P}/\mathrm{d}\boldsymbol{F}$

state::AbstractMaterialState: The updated material state variables at the end of the time increment

If the second input is the small strain tensor, $\boldsymbol{\epsilon}$ (ϵ::SymmetricTensor{2}), the outputs are

σ::SymmetricTensor{2}: Cauchy stress tensor, $\boldsymbol{\sigma}$

dσdϵ::SymmetricTensor{4}: Algorithmic tangent stiffness tensor, $\mathrm{d}\boldsymbol{\sigma}/\mathrm{d}\boldsymbol{\epsilon}$

state::AbstractMaterialState: The updated material state variables at the end of the time increment

If the second input is deformation vector, $\boldsymbol{u}$ (u::Vec), the outputs are

t::Vec: Traction vector, $\boldsymbol{t}$

dtdu::SecondOrderTensor: Algorithmic tangent stiffness tensor, $\mathrm{d}\boldsymbol{t}/\mathrm{d}\boldsymbol{u}$

state::AbstractMaterialState: The updated material state variables at the end of the time increment

Additionally, the following types and functions could be defined for a material

MaterialModelsBase.initial_material_state — Function

initial_material_state(m::AbstractMaterial)

Return the (default) initial state::AbstractMaterialState of the material m.

Defaults to the empty NoMaterialState()

MaterialModelsBase.get_cache — Function

get_cache(m::AbstractMaterial)

Return a cache::AbstractMaterialCache that can be used when calling the material to reduce allocations

Defaults to the empty NoMaterialCache()

MaterialModelsBase.AbstractExtraOutput — Type

AbstractExtraOutput()

By allocating an AbstractExtraOutput type, this type can be mutated to extract additional information from the internal calculations in material_response only in cases when this is desired. E.g., when calculating derivatives. The concrete NoExtraOutput<:AbstractExtraOutput exists for the case when no additional output should be calculated.