MaterialModelsBase

The main purpose of this package is to provide a unifying interface, facilitating the interchanging of material models between researchers. The basic idea, is to write packages (e.g. finite element simulations), that rely only on MaterialModelsBase.jl. It will then be easy to run different material models written by other researchers, to check if your implementation is correct or to evaluate different material models.

Basic usage

This section describes how to use a material model defined according to the described interface. For this example, we will use the case TestMaterials module used for testing purposes (see test/TestMaterials.jl in the main repo). Hence, replace TestMaterials with the specific material model package(s) you wish to use.

Use the required modules and define your material according to the specification

using MaterialModelsBase, TestMaterials, Tensors
G1=80.e3; G2=8.e3; K1=160.e3; K2=16.e3; η=1000.0
material = ViscoElastic(LinearElastic(G1,K1), LinearElastic(G2,K2), η)

As one who will run a code (e.g. a finite element solver or a material point simulator) that someone else have written, this is all that needs to be done. If we would write a material point simulator, we also need to instantiate the material state, and optionally the cache.

state = initial_material_state(material)
cache = get_cache(material)

Now, we can loop through the time history, let's say we want to simulate the uniaxial stress case:

stress_state = UniaxialStress() 
ϵ11_vector = collect(range(0, 0.01; length=100))   # Linear ramp
Δt = 1.e-3                                  # Fixed time step
for ϵ11 in ϵ11_vector
    ϵ11_tensor = SymmetricTensor{2,1}(ϵ11)
    σ, dσdϵ, state, ϵ_full = material_response(stress_state, material, ϵ11_tensor, state, Δt, cache)
end

where σ and dσdϵ are the reduced outputs for the given stress_state (in this case dimension 1). The state is not directly affected by the stress_state, and the ϵ_full gives the full strain tensor, such that

σ, dσdϵ, state = material_response(material, ϵ_full, old_state, Δt, cache)

gives the full stress and stiffness, adhering to the constraint imposed by the stress_state. Note that in the latter function call, no stress_state was given, and then no ϵ_full output was given either.

If the stress iterations would not converge, a NoStressConvergence<:MaterialConvergenceError exception is thrown. The package also contains the exception NoLocalConvergence<:MaterialConvergenceError, which should be thrown from inside implemented material routines to signal that something didn't converge and that the caller should consider to e.g. reduce the time step or handel the issue in some other way.

Advanced use cases

The package also provides an interface for converting AbstractMaterials to a vector of material parameters and vice versa. However, to implement this part of the interface is not required, if only the standard usage is intended for your material model.

Building upon that option, it is also possible to define differentiation of the material response wrt. to the material parameters, which is useful for parameter identification using gradient based methods. For further details, see Differentation of a material.

API

The main function is the material_response function that primarly dispatches on the AbstractMaterial input type.

MaterialModelsBase.material_response — Function

material_response(
    [stress_state::AbstractStressState],
    m::AbstractMaterial, 
    strain::Union{SecondOrderTensor,Vec}, 
    old::AbstractMaterialState, 
    Δt::Union{Number,Nothing}=nothing, 
    cache::AbstractMaterialCache=get_cache(m), 
    extras::AbstractExtraOutput=NoExtraOutput(); 
    options::Dict=Dict{Symbol}()
    )

Calculate the stress/traction, stiffness and updated state variables for the material m, given the strain input strain.

Mandatory arguments

m: Defines the material and its parameters

The second strain input can be, e.g.

ϵ::SymmetricTensor{2}: The total small strain tensor at the end of the increment

F::Tensor{2}: The total deformation gradient at the end of the increment

u::Vec: The deformation at the end of the increment (for cohesive elements)

old::AbstractMaterialState: The material state variables at the end of the last converged increment. The material initial material state can be obtained by the initial_material_state function.

Optional positional arguments

stress_state: Use to solve for a reduced stress state, e.g. PlaneStress. See Stress states.

Δt: The time step in the current increment. Defaults to nothing.

cache::AbstractMaterialCache: Cache variables that can be used to avoid allocations during each call to the material_response function. This can be created by the get_cache function.

extras: Updated with requested extra output. Defaults to the empty struct NoExtraOutput

Optional keyword arguments

options: Additional options that may be specific for each material. This is also used for stress iterations, see Stress states.

Outputs

stress, is the stress measure that is energy conjugated to the strain (2nd) input.

stiffness, is the derivative of the stress output wrt. the strain input.

new_state, are the updated state variables

strain, is only available if stress_state is given and returns the full strain tensor

The following are probably the three most common strain and stress pairs:

If the second input is the deformation gradient $\boldsymbol{F}$ (F::Tensor{2})`, the outputs are

P::Tensor{2}: First Piola-Kirchhoff stress, $\boldsymbol{P}$

dPdF::Tensor{4}: Algorithmic tangent stiffness tensor, $\mathrm{d}\boldsymbol{P}/\mathrm{d}\boldsymbol{F}$

If the second input is the small strain tensor, $\boldsymbol{\epsilon}$ (ϵ::SymmetricTensor{2}), the outputs are

σ::SymmetricTensor{2}: Cauchy stress tensor, $\boldsymbol{\sigma}$

dσdϵ::SymmetricTensor{4}: Algorithmic tangent stiffness tensor, $\mathrm{d}\boldsymbol{\sigma}/\mathrm{d}\boldsymbol{\epsilon}$

If the second input is deformation vector, $\boldsymbol{u}$ (u::Vec), the outputs are

t::Vec: Traction vector, $\boldsymbol{t}$

dtdu::SecondOrderTensor: Algorithmic tangent stiffness tensor, $\mathrm{d}\boldsymbol{t}/\mathrm{d}\boldsymbol{u}$

Additionally, the following types and functions could be defined for a material

MaterialModelsBase.initial_material_state — Function

initial_material_state(m::AbstractMaterial)

Return the (default) initial state::AbstractMaterialState of the material m.

Defaults to the empty NoMaterialState()

MaterialModelsBase.get_cache — Function

get_cache(m::AbstractMaterial)

Return a cache::AbstractMaterialCache that can be used when calling the material to reduce allocations

Defaults to the empty NoMaterialCache()

MaterialModelsBase.AbstractExtraOutput — Type

AbstractExtraOutput()

By allocating an AbstractExtraOutput type, this type can be mutated to extract additional information from the internal calculations in material_response only in cases when this is desired. E.g., when calculating derivatives or for multiphysics simulations. The concrete NoExtraOutput<:AbstractExtraOutput exists for the case when no additional output should be calculated.

Finally, the following exceptions are included

MaterialModelsBase.MaterialConvergenceError — Type

MaterialConvergenceError

Can be used to catch errors related to the material not converging.

MaterialModelsBase.NoLocalConvergence — Type

NoLocalConvergence(msg::String)

Throw if the material_response routine doesn't converge internally

MaterialModelsBase.NoStressConvergence — Type

NoStressConvergence(msg::String)

This is thrown if the stress iterations don't converge, see Stress states