Differentation of a material

MaterialModelsBase.getnumtensorcomponents — Function

getnumtensorcomponents(::AbstractMaterial)

Returns the number of independent components for the given material.

If the material works with the small strain tensor and Cauchy stress, return 6 (default)

If the material works with the deformation gradient and the 1st Piola-Kirchhoff stress, return 9

If the material is a cohesive material working with vectors, return the number of vector components (e.g. 3)

Defaults to 6 if not overloaded

MaterialModelsBase.getnumstatevars — Function

getnumstatevars(m::AbstractMaterial)

Return the number of state variables. A tensorial state variable should be counted by how many components it has. E.g. if a state consists of one scalar and one symmetric 2nd order tensor, getnumstatevars should return 7 (if the space dimension is 3).

Defaults to 0 if not overloaded

MaterialModelsBase.getnumparams — Function

getnumparams(m::AbstractMaterial)

Return the number of material parameters in m. No default value implemented.

MaterialModelsBase.material2vector! — Function

material2vector!(v::AbstractVector, m::AbstractMaterial)

Put the material parameters of m into the vector m. This is typically used when the parameters should be fitted.

MaterialModelsBase.vector2material — Function

vector2material(v::AbstractVector, ::MT) where {MT<:AbstractMaterial}

Create a material of type MT with the parameters according to v

MaterialModelsBase.material2vector — Function

material2vector(m::AbstractMaterial)

Out-of place version of material2vector!. Given getnumparams, this function does not need to be overloaded unless another datatype than Float64 should be used.

MaterialModelsBase.MaterialDerivatives — Type

MaterialDerivatives(m::AbstractMaterial, T=Float64)

A struct that saves all derivative information using a Matrix{T} for each derivative. If getnumtensorcomponents, getnumstatevars, and getnumparams are implemented for m, MaterialDerivatives does not need to be overloaded for m.

MaterialModelsBase.allocate_differentiation_output — Function

allocate_differentiation_output(::AbstractMaterial)

When calculating the derivatives of a material, it can often be advantageous to have additional information from the solution procedure inside material_response. This can be obtained via an AbstractExtraOutput, and allocate_differentiation_output provides a standard function name for what extra_output::AbstractExtraOutput that should be allocated in such cases.

Defaults to an NoExtraOutput if not overloaded.

MaterialModelsBase.differentiate_material! — Function

differentiate_material!(
    diff::MaterialDerivatives, 
    m::AbstractMaterial, 
    ϵ::Union{SecondOrderTensor, Vec}, 
    old::AbstractMaterialState, 
    Δt,
    cache::AbstractMaterialCache
    dσdϵ::AbstractTensor, 
    extra::AbstractExtraOutput;
    options=nothing
    )

Calculate the derivatives and save them in diff.